4_financial_analyst_mcp/requirements.txt
fastmcp
crewai
crewai-tools
google-generativeai
langchain-google-genai
yfinance
pandas
matplotlib
python-dotenv
pydantic
streamlit
